Implemented feature: Randomized availability of items in shop cells
Implemented a feature in the shop where the availability of items in each shop cell is randomized. This adds variability to the gameplay experience, as some cells may be empty while others contain items for purchase. Players will encounter different configurations of available items in each playthrough, adding an element of surprise and strategy to their decision-making in the shop.